backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.32
pydantic>=2.0.0
pydantic-settings==2.4.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
aiofiles==24.1.0
alembic==1.13.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
